requirements.txt
bcrypt==3.1.7
cffi==1.12.3
mysql-connector-python==8.0.17
mysqlclient==1.4.4
protobuf==3.10.0
pycparser==2.19
python-dotenv==0.10.3
six==1.12.0
